+++ src/containers/Manufacturer/index.module.scss
@@ -10,11 +10,4 @@
     justify-content: center;
     margin-left: 30px;
   }
-
-  .Field {
-    background: #141414;
-    border: 1px solid #434343;
-    border-radius: 5px;
-    margin-bottom: 5px;
-  }
 }
