--- src/containers/System/components/Firmware/index.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Firmware = () => <h1>Firmware Page</h1>;
-
-export default Firmware;
