+++ src/index.js
@@ -12,6 +12,8 @@ export { default as AddProfile } from 'containers/AddProfile';
 export { default as Accounts } from 'containers/Accounts';
 export { default as Alarms } from 'containers/Alarms';
 export { default as System } from 'containers/System';
+export { default as Manufacturer } from 'containers/Manufacturer';
+
 export { default as ClientDeviceDetails } from 'containers/ClientDeviceDetails';
 
 export { default as ThemeProvider } from 'contexts/ThemeProvider';
